feature: created Login hook and save and recover user in sesion storage
