+++ src/components/login/Login.tsx
@@ -1,88 +1,14 @@
-import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import LoginRegister from 'services/LoginRegister';
-import { toast } from 'sonner';
-import { useUserStore } from 'store/userStore';
 import { btnDefault } from 'styles/tailwind.classes';
-import { LoginRequest, User } from 'types/apiTypes';
-import { validateForm } from 'utils/validations';
 import InputBase from '../InputBase';
-
-interface ErrorsForm {
-  username?: string;
-  password?: string;
-  error?: string;
-}
+import { useLogin } from 'hooks/useLogin';
 
 interface LoginProps {
   handleLogin: () => void;
 }
+
 export function Login({ handleLogin }: LoginProps) {
-  const setUser = useUserStore((state) => state.setUser);
-  const setTokenStore = useUserStore((state) => state.setToken);
-  const setIsLogged = useUserStore((state) => state.setIsLogged);
-  const setRefreshToken = useUserStore((state) => state.setRefreshToken);
-
-  const navigate = useNavigate();
-  const [infoUser, setInfoUser] = useState({
-    username: '',
-    password: ''
-  });
-  const [errors, setErrors] = useState<ErrorsForm>({});
-
-  const sendData = async (infoUser: LoginRequest) => {
-    try {
-      const errorsForm: ErrorsForm = validateForm([
-        { name: 'userName', value: infoUser.username, required: true },
-        {
-          name: 'password',
-          value: infoUser.password,
-          required: true,
-          minLength: 8
-        }
-      ]);
-
-      setErrors(errorsForm);
-      if (!errorsForm.username && !errorsForm.password) {
-        await loginUser(infoUser);
-        navigate('/');
-      }
-    } catch (err) {
-      if (err instanceof Error) setErrors({ error: err.message });
-    }
-  };
-
-  const handleSignupClick = (
-    e: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ) => {
-    e.preventDefault();
-    handleLogin();
-  };
-
-  const loginUser = async (credentials: LoginRequest): Promise<void> => {
-    const response = await LoginRegister.login(credentials);
-    const { token, user, refreshToken } = response;
-    if (token) {
-      localStorage.setItem('token', JSON.stringify(token)); // Store the user token in local storage
-      saveGlobalStateUser(token, user, refreshToken);
-      toast.success('Login successful', {
-        duration: 2000
-      });
-    } else {
-      throw new Error('Incorrect username or password');
-    }
-  };
-
-  const saveGlobalStateUser = (
-    token: string,
-    user: User,
-    refreshToken: string
-  ) => {
-    setTokenStore(token);
-    setUser(user);
-    setRefreshToken(refreshToken);
-    setIsLogged(true);
-  };
+  const { infoUser, errors, sendData, handleSignupClick, setInfoUser } =
+    useLogin({ handleLogin });
 
   return (
     <div className="flex flex-col justify-center items-center p-0 m-0 ">
